requirements.txt
pyautogui
Pillow
pytesseract
pyyaml
opencv-python
pytest
